% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_site.R
\name{clear_site}
\alias{clear_site}
\title{Clear all files in the site directory}
\usage{
clear_site(dir = "./docs")
}
\arguments{
\item{dir}{directory to wipe out, defaults to ./docs per Github Pages convention}
}
\description{
Clear all files in the site directory
}
