% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_info_site.R
\name{create_info_site}
\alias{create_info_site}
\title{Create infographics web site}
\usage{
create_info_site(site_title = "Florida Keys Infographics", path_root = ".",
  dir_svg = "svg", elements_csv = "svg_elements.csv",
  indicators_csv = "plot_indicators.csv", dir_rmd = "rmd",
  dir_web = "docs", svg_paths = list.files(file.path(path_root, dir_svg),
  ".*\\\\.svg$", full.names = T),
  svg_names = tools::file_path_sans_ext(basename(svg_paths)),
  site_yml_brew = system.file("site_template/_site.yml.brew", package =
  "infographiq"), index_md_brew = system.file("site_template/index.md.brew",
  package = "infographiq"),
  readme_md_brew = system.file("site_template/README.md.brew", package =
  "infographiq"), header = system.file("site_template/_header.html", package =
  "infographiq"), footer = system.file("site_template/_footer.html", package =
  "infographiq"), styles_css = system.file("site_template/styles.css", package
  = "infographiq"), index = system.file("site_template/index.Rmd", package =
  "infographiq"), render_modals = T)
}
\arguments{
\item{site_title}{}

\item{path_root}{}

\item{dir_svg}{}

\item{elements_csv}{}

\item{indicators_csv}{csv table containing pointers going from svg_id to csv_url with data for plot of indicator...}

\item{dir_rmd}{}

\item{dir_web}{}

\item{svg_paths}{}

\item{svg_names}{}

\item{site_yml_brew}{}

\item{index_md_brew}{}

\item{readme_md_brew}{}

\item{header}{}

\item{footer}{}

\item{styles_css}{}

\item{index}{}
}
\value{
nothing returned from the function if no error, except the created site is generated. http://rmarkdown.rstudio.com/rmarkdown_websites.html
}
\description{
Create infographics web site
}
\examples{
library(infographiq)
create_info_site()
}
