% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{create_website}
\alias{create_website}
\title{Create new site from template}
\usage{
create_website(dir_path, open = rlang::is_interactive())
}
\arguments{
\item{dir_path}{The directory path where the new site should be located.}
}
\value{
The output is a directory containing the template version of a site.
}
\description{
This function creates a new website from a template stored within the package.
}
\examples{
\dontrun{
directory_name("test_location")
}

}
