% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{get_modal_info}
\alias{get_modal_info}
\title{Generate introductory info for the html of a modal window}
\usage{
get_modal_info(
  rmd = knitr::current_input(),
 
    info_modal_links_csv = "https://docs.google.com/spreadsheets/d/1yEuI7BT9fJEcGAFNPM0mCq16nFsbn0b-bNirYPU5W8c/gviz/tq?tqx=out:csv&sheet=info_modal_links"
)
}
\arguments{
\item{rmd}{The name of an input file passed to knit().}

\item{info_modal_links_csv}{A hyperlink to the google sheet, in csv format, that contains the modal links info.}
}
\value{
The function returns a string that is a set of html tags to be inserted into a html file.
}
\description{
This function generates the html tags for the top portion of a rmd modal window,
containing the introductory information about that window. This function only
will work within a rmd file to be knitted.
}
\examples{
\dontrun{
get_modal_info()
}

}
