% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_plotting_function_brew.R
\name{get_plotting_function_brew}
\alias{get_plotting_function_brew}
\title{takes a string value (assumed to be the value of the plotting_function column
from plot_indicators.csv) and returns the brew template we think the user
wants.}
\usage{
get_plotting_function_brew(plotting_function_call)
}
\arguments{
\item{plotting_function_call}{str from plot_indicators.csv$plotting_function}
}
\value{
returns brew file object that can be passed to brew()
}
\description{
takes a string value (assumed to be the value of the plotting_function column
from plot_indicators.csv) and returns the brew template we think the user
wants.
}
\examples{
 plot_brew = get_plotting_function_brew(
}
