% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{glossarize_md}
\alias{glossarize_md}
\title{Insert html tags for glossary tooltips into md files}
\usage{
glossarize_md(md, md_out = md)
}
\arguments{
\item{md}{The markdown file where the tags are to be inserted.}

\item{md_out}{The markdown output file.}
}
\value{
The output of this function is html tags inserted into a markdown file.
}
\description{
The purpose of this function is to insert the html tags required for glossary
tooltip functionality into a given markdown file. The function glossarize_md
is used by the function rmd2html described in this package.
}
