% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{infographiq_template}
\alias{infographiq_template}
\title{Create new site from template}
\usage{
infographiq_template(path)
}
\arguments{
\item{path}{The directory path where the new site should be located.}
}
\value{
The output is a directory containing the template version of a site.
}
\description{
This function creates a new website from a template stored within the package.
}
\examples{
\dontrun{
infographiq_template("test_location")
}

}
