% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{insert_tooltip}
\alias{insert_tooltip}
\title{Insert tooltips into text.}
\usage{
insert_tooltip(text, glossary_term, span_css)
}
\arguments{
\item{text}{The section of text where tooltips are to be added.}

\item{glossary_term}{The glossary term to be looked for.}

\item{span_css}{The css tags to add before the glossary term.}
}
\value{
The function outputs a string containing the text section with html tags inserted.
}
\description{
The purpose of the following function is, for a provided section of text, to
insert the required tooltip css around a provided glossary term. The function
preserves the pattern of capitalization of the glossary term that already exists.
This function is used by the function glossarize_md also described in this
package.
}
