% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{md_caption}
\alias{md_caption}
\title{Generate the caption for a figure}
\usage{
md_caption(
  title,
  md = here::here("modals/_captions.md"),
  get_details = F,
  fig_in_report = T
)
}
\arguments{
\item{title}{The name of a figure.}

\item{md}{The md file containing the list of captions.}

\item{get_details}{A Boolean variable indicating whether a short or expanded caption is required.}

\item{fig_in_report}{A Boolean variable indicating whether the figure described in the caption is present in the condition report.}
}
\value{
A string containing the caption, with html tags inserted.
}
\description{
This function generates either a short or expanded caption for a given figure.
}
\examples{
\dontrun{
md_caption("Figure Ux.Ocean.SST.ERD.map.", get_details = T)
}

}
