% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svg.R
\name{modal_html}
\alias{modal_html}
\title{Get modal html}
\usage{
modal_html(modal_id = "modal")
}
\arguments{
\item{modal_id}{id of div tag to use for modal window}
}
\value{
HTML tag
}
\description{
Get modal html
}
\examples{
modal_html()
}
