% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ocnms_render_rmd.R
\name{ocnms_get_modal_info}
\alias{ocnms_get_modal_info}
\title{OCNMS version: generate introductory info for the html of a modal window}
\usage{
ocnms_get_modal_info(rmd = knitr::current_input())
}
\arguments{
\item{rmd}{The name of an input file passed to knit().}
}
\value{
The function returns a string that is a set of html tags to be inserted into a html file.
}
\description{
This function generates the html tags for the top portion of a rmd modal window,
containing the introductory information about that window. This function only
will work within a rmd file to be knitted.
}
\examples{
\dontrun{
ocnms_get_modal_info()
}

}
