% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_dygraph_timeseries}
\alias{plot_dygraph_timeseries}
\title{plot timeseries given a path (or URL) to csv}
\usage{
plot_dygraph_timeseries(
  csv_tv,
  title,
  y_label,
  x_label = "Year",
  v_label = y_label,
  filter = NULL,
  col_t = NULL,
  col_y = NULL,
  skip = 0,
  use_kmb = T,
  group_by = NULL,
  std_err = NULL
)
}
\arguments{
\item{csv_tv}{csv path or URL with time and value}

\item{title}{title of plot}

\item{y_label}{label (eg units) for values on y axis}

\item{x_label}{label for date/time on x axis}

\item{v_label}{label for legend, defaults to y_label}

\item{filter}{filter expression to limit rows}

\item{col_t}{column name for time}

\item{col_y}{column name for y}

\item{skip}{rows to skip in csv header}
}
\value{
returns dygraph htmlwidget for populating modal window
}
\description{
plot timeseries given a path (or URL) to csv
}
\examples{
plot_dygraph_timeseries(
  csv_tv  = 'http://oceanview.pfeg.noaa.gov/erddap/tabledap/cciea_MM_pup_count.csv?time,mean_growth_rate',
  title   = 'Female sea lion pup growth rate',
  y_label = 'Mean growth rate')
}
