% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_intertidal_nms}
\alias{plot_intertidal_nms}
\title{Plot Intertidal Timeseries for National Marine Sanctuaries}
\usage{
plot_intertidal_nms(d_csv, nms, sp, sp_name)
}
\arguments{
\item{d_csv}{csv file of timeseries data containing fields: site, date, pct_cover}

\item{nms}{abbreviation for national marine sanctuary, also included as a site averaged over others}

\item{sp}{species code}

\item{sp_name}{species name}
}
\value{
timeseries plot
}
\description{
Plot Intertidal Timeseries for National Marine Sanctuaries
}
