% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_all_rmd}
\alias{render_all_rmd}
\title{Render all rmd files in the modals folder that have been changed}
\usage{
render_all_rmd(nms = "cinms", interactive_only = F, render_all = F)
}
\arguments{
\item{nms}{The NMS sanctuary with only "cinms" currently doing anything.}

\item{interactive_only}{A Boolean variable indicating whether only rmd files containing interactive figures should be rendered. NOTE: If this is set to "TRUE", the rmd files containing interactive figures with MARINe data will be omitted. These latter figures can only be rendered in a local environment as they require connection to a shared Google Drive that is accessed via Google Drive File Stream.}

\item{render_all}{A Boolean variable indicating whether all rmd files should be rendered, whether or not there have been changes to them.}
}
\value{
The function outputs a html file for every rmd file in the modals folder.
}
\description{
Render all rmd files in the modals folder that have been changed
}
\examples{
\dontrun{
render_all_rmd(interactive_only = T)
}

}
