% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{render_figure}
\alias{render_figure}
\title{Produce full html for static figures, minus tooltips.}
\usage{
render_figure(figure_id, figure_img)
}
\arguments{
\item{figure_id}{The id of the figure.}

\item{figure_img}{The path of the figure image.}
}
\value{
The output is a string containing the html tags to display the figure and figure caption.
}
\description{
This is a function that generates the html to display a static figure and the
captions for that figure. Glossary tooltips are not created here, as that
occurs at a later stage of the html production process.
}
\examples{
\dontrun{
render_figure("Figure App.C.4.4.", "../img/cinms_cr/App.C.4.4.Leeworthy_landings.jpg")
}

}
