% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_rmd.R
\name{rmd2html}
\alias{rmd2html}
\title{Render html for rmd file, including glossary tooltips.}
\usage{
rmd2html(rmd)
}
\arguments{
\item{rmd}{The R markdown file to be rendered into html.}
}
\value{
The output is a html file that is the rendered rmd file.
}
\description{
This function creates all of the html for a R markdown file, inserting in the
glossary tooltips. Note that this function will break html that contains interactive
figures, as most of the required javascript won't be loaded into the html <head>.
}
\examples{
\dontrun{
rmd2html(here::here("modals/ca-sheephead.Rmd"))
}
}
